Finally figured out how to rewrite everything to use requirejs!
Engine namespace, temporary gameloop, all modules, and all module substructures are using define to properly get dependencies and load each module in the proper order.

Things are also much more abstracted as a result, and the engine is no longer defined as a public namespace (kinda a nice addition, though still working out debugging issues).

Currently using the requirejs/jquery mashup, so it's uncompressed. Luckily, this won't be an issue, since part of the move to requirejs was to allow for code compilation and minification.

Basically, new pattern has be established properly across all current files, and we can now work on streamlining the pattern and get back to actually writing code.