Reconfigured project to be only the engine. Before the structure was as if there was going to be a game in the same repo. This is not how we want our modular engine setup.
Added a config system. Modules are managed in config.js. It is simply a JSON object that has a modules object, containing the engine modules you wish to load,
and each of those modules contains an array of substructures you wish to load with them. This JSON object also has two built-in functions, module_paths and substructure_paths.
These are designed to parse the modules and return proper paths for them all.

The whole config system is really hacked together, and will need some serious refining, but I think it's a step in the right direction.

Also of note is that to load the config file, I'm using a requirejs shim. This should go asap, especially since config is basically a globally accessible list right now.
We need to find a good way to make config an AMD module and load it as a dependency of the engine. I just haven't figured out how yet.