- Fleshed out props a bit more. Edited sprites to keep track of their current position/size in the spritesheet and the image they reference. - Updated test loop to reflect the changes. - Render methods now take a prop as an argument
The result of these changes is that the prop can contain everything the renderer needs to know to draw that prop.

We also got a massive performance boost, though I'm not sure what I changed that caused it. Might have something to do with moving
the requestAnimationFrame call to the beginning of the update call, instead of the end. Anyways, down from 25% CPU (100% of a thread)
to about 5% CPU, which also means it's rendering consistently at 62 fps.