Added attach_engine to config.js
This allows the engine to be accessed freely in any modules or submodules.

From now on, anytime you need to access something within the engine namespace from
a module or submodule, you can use this.engine to reference it. Note that
this.engine is added after the object definition for a module or submodule.