Updates to error module, resource module, and parser module.
- Updated error module to accept a standard Error object for each error method
- Updated parser module to include basic skeleton. (WARNING: Still looking at possible layouts. We may not even need parser)
- Updated resource module to have more specific methods and fleshed out the addResource method (should actually work now)

Next on the list is getting actual resource files to load, and getting configuration files to load,
and then be automatically parsed by the engine to grab resources and set themselves up. We then need
to get unload and remove resource methods working, so we can test how well we can remove unneeded resources from memory.